requirements.txt
flask
google-generativeai
Pillow
gunicorn